miner separted from nucleus
